% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{resultC}
\alias{resultC}
\title{Bias of estimated sigma and average model size using Rcpp}
\usage{
resultC(hsigma, hbeta)
}
\arguments{
\item{hsigma}{the estimated sigma}

\item{hbeta}{the estimated coefficients}
}
\value{
bias of estimated sigma and average model size
}
\description{
Bias of estimated sigma and average model size using Rcpp
}
