\name{SC19027-package}
\alias{SC19027-package}
\docType{package}
\title{
  The R package for StatComp Course by 19027
}
\description{
 The package includes a data generation function and linear regression function for scaled MCP or SCAD.It also includes all my homework answers.
}
\author{
Chen Shuyue

Maintainer: Chen Shuyue <csy619@mail.ustc.edu.cn>
}
\keyword{ package }