% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalescad_mcp.R
\name{scalescad_mcp}
\alias{scalescad_mcp}
\title{The Scaled MCP or SCAD}
\usage{
scalescad_mcp(X, y, lam0, method)
}
\arguments{
\item{X}{X is an n by p design matrix where n is the sample size and p is the data dimension.}

\item{y}{Y is the n dimensional response vector.}

\item{lam0}{the initial value of penalty level.}

\item{method}{the penalized methods,can be "scad" or "mcp".}
}
\value{
hsigma the estimated noise level

coefficients the estimated coefficients
}
\description{
The scaled penalized methods with the minimax concave penalty(MCP) or smoothly clipped absolute deviation penalty(SCAD).
}
\examples{
\dontrun{
n<-200;p<-2000
data<-datageneration(n,p,0)
X<-data$X
Y<-data$Y
lam0<-sqrt(2*log(p)/n)
mcp<-scalescad_mcp(X,Y,lam0,method="mcp")
sigma<-mcp$hsigma
}
}
